+++ src/test/ExSln2/c99/gen/mcu/stm32/Stm32Gpio.h
@@ -49,13 +49,25 @@ void mcu_stm32_Stm32Gpio_toggle(mcu_stm32_Stm32Gpio * self);
 extern const hal_IGpio_vtable mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp;
 
 // Up conversion from mcu_stm32_Stm32Gpio to hal_IGpio interface
-#define M_mcu_stm32_Stm32Gpio__to__hal_IGpio(self_arg)    (hal_IGpio){ .obj = self_arg, .obj_vtable = (const hal_IGpio_vtable*)(&mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp.read_state) }
+// MAA stands for Macro Aggregate Assignment. See https://github.com/fin-language/fin/issues/60 
+#define MAA_mcu_stm32_Stm32Gpio__to__hal_IGpio(self_arg)    { .obj = self_arg, .obj_vtable = (const hal_IGpio_vtable*)(&mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp.read_state) }
+// MCL stands for Macro Compound Literal. See https://github.com/fin-language/fin/issues/60 
+#define MCL_mcu_stm32_Stm32Gpio__to__hal_IGpio(self_arg)    (hal_IGpio){ .obj = self_arg, .obj_vtable = (const hal_IGpio_vtable*)(&mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp.read_state) }
 
 // Up conversion from mcu_stm32_Stm32Gpio to hal_IDigInOut interface
-#define M_mcu_stm32_Stm32Gpio__to__hal_IDigInOut(self_arg)    (hal_IDigInOut){ .obj = self_arg, .obj_vtable = (const hal_IDigInOut_vtable*)(&mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp.read_state) }
+// MAA stands for Macro Aggregate Assignment. See https://github.com/fin-language/fin/issues/60 
+#define MAA_mcu_stm32_Stm32Gpio__to__hal_IDigInOut(self_arg)    { .obj = self_arg, .obj_vtable = (const hal_IDigInOut_vtable*)(&mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp.read_state) }
+// MCL stands for Macro Compound Literal. See https://github.com/fin-language/fin/issues/60 
+#define MCL_mcu_stm32_Stm32Gpio__to__hal_IDigInOut(self_arg)    (hal_IDigInOut){ .obj = self_arg, .obj_vtable = (const hal_IDigInOut_vtable*)(&mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp.read_state) }
 
 // Up conversion from mcu_stm32_Stm32Gpio to hal_IDigIn interface
-#define M_mcu_stm32_Stm32Gpio__to__hal_IDigIn(self_arg)    (hal_IDigIn){ .obj = self_arg, .obj_vtable = (const hal_IDigIn_vtable*)(&mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp.read_state) }
+// MAA stands for Macro Aggregate Assignment. See https://github.com/fin-language/fin/issues/60 
+#define MAA_mcu_stm32_Stm32Gpio__to__hal_IDigIn(self_arg)    { .obj = self_arg, .obj_vtable = (const hal_IDigIn_vtable*)(&mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp.read_state) }
+// MCL stands for Macro Compound Literal. See https://github.com/fin-language/fin/issues/60 
+#define MCL_mcu_stm32_Stm32Gpio__to__hal_IDigIn(self_arg)    (hal_IDigIn){ .obj = self_arg, .obj_vtable = (const hal_IDigIn_vtable*)(&mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp.read_state) }
 
 // Up conversion from mcu_stm32_Stm32Gpio to hal_IDigOut interface
-#define M_mcu_stm32_Stm32Gpio__to__hal_IDigOut(self_arg)    (hal_IDigOut){ .obj = self_arg, .obj_vtable = (const hal_IDigOut_vtable*)(&mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp.set_state) }
+// MAA stands for Macro Aggregate Assignment. See https://github.com/fin-language/fin/issues/60 
+#define MAA_mcu_stm32_Stm32Gpio__to__hal_IDigOut(self_arg)    { .obj = self_arg, .obj_vtable = (const hal_IDigOut_vtable*)(&mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp.set_state) }
+// MCL stands for Macro Compound Literal. See https://github.com/fin-language/fin/issues/60 
+#define MCL_mcu_stm32_Stm32Gpio__to__hal_IDigOut(self_arg)    (hal_IDigOut){ .obj = self_arg, .obj_vtable = (const hal_IDigOut_vtable*)(&mcu_stm32_Stm32Gpio__hal_IGpio_vtable_imp.set_state) }